+++ NEWS.md
@@ -1,3 +1,9 @@
+# imcover 0.2.5
+
+* Fix linking of Stan code and rebuilding models.
+* Update links to download input datasets.
+* Handle alternate survey evidence types.
+
 # imcover 0.2.4
 
 * Updated documentation